Crate lea_vm

This crate provides the VM used to execute compiled Lea programs.

Since this is considered an internal crate, everything inside here is entirely unstable and unsafe (because Rust really doesn't like it when I strap a GC on top of it).

Modules

error
Runtime errors

function
Defines the FunctionProto type, which is a garbage-collected version of core's FnData, as well as the Function type, which is an instantiated, runnable function inside the VM.

libfn
Types and macros for defining library functions in Rust, callable from Lea.

mem

number

Macros

lea_lib!
Generate a pub fn init which will populate a Table. This can be used to declare loadable libraries.

lea_libfn!
Declare a list of library functions callable from Lea code.

Structs

Array
An array that can contain Lea values. This represents the Array value.

Str
Str is an immutable String.

Table
A table that maps Lea values to Lea values.

VM
A VM context. Holds a garbage collector that manages the program's memory, the stack used for local and temporary variables, the callstack, etc.

Enums

Value
The type used to represent all Lea values.