Enum lea_vm::Value

pub enum Value {
    Nil,
    Bool(bool),
    Number(Number),
    String(TracedRef<Str>),
    Closure(TracedRef<Function>),
    Array(TracedRef<Array>),
    Table(TracedRef<Table>),
    LibFn(LibFn),
}

The type used to represent all Lea values.

Variants
NilBool(bool)Number(Number)String(TracedRef<Str>)Closure(TracedRef<Function>)Array(TracedRef<Array>)Table(TracedRef<Table>)LibFn(LibFn)
Methods

impl Value

fn get_type_name(&self) -> &'static str

The equivalent of Lua's type() function, returns the type name of a value.

fn is_truthy(&self) -> bool

Is this value considered true or false when used in a boolean context (if/loop condition, shortcircuit operators)?

fn trace<T: Tracer>(&self, t: &mut T)

If this Value references a GC-object, marks it using the given Tracer.

fn from_literal<G: GcStrategy>(c: Const, gc: &mut G) -> Value

unsafe fn fmt<G: GcStrategy, W: Write>(&self, f: W, gc: &G) -> Result<()>

Formats this Value. This requires a reference to the GC that owns the referenced objects and is unsafe because the TracedRefs are dereferenced using the GC.