Struct lea_vm::function::FunctionProto

pub struct FunctionProto {
    pub source_name: String,
    pub stacksize: u8,
    pub params: u8,
    pub varargs: bool,
    pub opcodes: Opcodes,
    pub consts: Vec<Value>,
    pub upvalues: Vec<UpvalDesc>,
    pub upval_names: Vec<String>,
    pub lines: Vec<usize>,
    pub child_protos: Vec<TracedRef<FunctionProto>>,
}

A compiled function. When instantiated by the VM, becomes a Function.

Fields
source_name: String

The name of the source from which this function was compiled

stacksize: u8

The number of stack slots required by this function (might be dynamically increased)

params: u8

Number of parameters accepted (ignoring varargs)

varargs: bool

True if declared as a varargs function

opcodes: Opcodes

The opcodes emitted for the code in the function bodyt.mark_traceable(r),

consts: Vec<Value>

Constants used by this function.

upvalues: Vec<UpvalDesc>

List of Upvalue reference descriptions

upval_names: Vec<String>

Names of Upvalues (names may not be defined)

lines: Vec<usize>

Contains the last opcode number emitted for a given line

child_protos: Vec<TracedRef<FunctionProto>>

References to the prototypes of all function declared within the body of this function. When dynamically compiling code, this allows the GC to collect prototypes that are unused.

Methods

impl FunctionProto

fn from_fndata<G: GcStrategy>(fndata: FnData, gc: &mut G) -> TracedRef<FunctionProto>