Module lea_vm::libfn

Types and macros for defining library functions in Rust, callable from Lea.

Structs

LibFn
Newtype of a function reference to be used for library functions.

Enums

LibFnError
Kinds of error objects a LibFn can return

TyMarker
Describes how an argument type of a function is matched.

Traits

ToValues
Trait for all types that can be used in the return [...] of lib functions.

Type Definitions

LibFnTyInfo
Records type information about a function. Currently unused, but might come in handy when autogenerating documentation or performing correctness analyses.