Enum lea_vm::libfn::TyMarker

pub enum TyMarker {
    Number,
    String,
    Bool,
    Array,
    Table,
    Fn,
    Any,
    Varargs,
}

Describes how an argument type of a function is matched.

Variants
Number

number

String

string

Bool

bool

Array

array

Table

table

Fn

fn

Any

*: Matches any value

Varargs

...: Matches a list of values with any type (or an empty list)