Trait lea_vm::mem::Tracer

pub trait Tracer {
    unsafe fn mark_untraceable<T>(&mut self, TracedRef<T>) where T: Any;
    fn mark_traceable<T>(&mut self, TracedRef<T>) where T: Traceable;
}

A Tracer provides methods to mark objects as reachable. A GcStrategy will pass an object that implements Tracer to the trace method of all Traceable and reachable objects.

Required Methods

unsafe fn mark_untraceable<T>(&mut self, TracedRef<T>) where T: Any

Marks a TracedRef as reachable without following the referenced object's own references (ie. without tracing it).

This is unsafe and may only be called if the object really doesn't carry any references (currently only strings) (FIXME: Needs either some hack to work safely or specialization).

fn mark_traceable<T>(&mut self, TracedRef<T>) where T: Traceable

Marks an object as reachable. This will also (at some undefined point in the future) trace the object, finding all transitively reachable objects.