Struct lea_vm::VM

pub struct VM {
    pub gc: DefaultGc,
    pub stack: Vec<Value>,
    // some fields omitted
}

A VM context. Holds a garbage collector that manages the program's memory, the stack used for local and temporary variables, the callstack, etc.

Fields
gc: DefaultGc stack: Vec<Value>

"VM stack", "value stack" or just stack. Stores the activation of functions in the form of variables (registers) used by the function.

Methods

impl VM

fn new() -> VM

fn start<F>(&mut self, f: TracedRef<Function>, e: F) -> Option<Vec<Value>> where F: FnOnce(&mut VmError)

Starts execution of the program. This may not be called while the VM is already executing a function.

Returns when the function returns (or a runtime error occurs).

impl VM

fn cur_func(&self) -> &Function

fn main_ref(&self) -> TracedRef<Function>

Obtains a TracedRef<Function> pointing at the main function (root of the call stack). Will panic when the VM isn't executing a function.